Remove overpay stop from socks contract
Explanation: The signing flow for Tezos transactions is such that a severe overpayment
almost has to be intentional. And if it isn't, doing refund flows/etc with smart
contracts is difficult. It's irresponsible to promote the expectation that you
can revert or prevent sloppy work from losing you money. Furthermore contracts
like this are typically going to be used from some kind of client. It is both
trivial and the proper responsibility of the client to prevent overpays, not
the 'server' or contract.